[expo-updates][ios][3/n] Remove classic updates
