+++ packages/expo-manifests/ios/EXManifests/ManifestFactory.swift
@@ -6,12 +6,10 @@ import Foundation
 @objcMembers
 public class ManifestFactory: NSObject {
   public static func manifest(forManifestJSON: [String: Any]) -> Manifest {
-    if forManifestJSON["releaseId"] != nil {
-      return LegacyManifest(rawManifestJSON: forManifestJSON)
-    } else if forManifestJSON["metadata"] != nil {
+    if forManifestJSON["metadata"] != nil {
       return NewManifest(rawManifestJSON: forManifestJSON)
-    } else {
-      return BareManifest(rawManifestJSON: forManifestJSON)
     }
+
+    return BareManifest(rawManifestJSON: forManifestJSON)
   }
 }
