+++ packages/expo-updates/ios/EXUpdates/Update/Update.swift
@@ -75,6 +75,7 @@ public enum UpdateStatus: Int {
 @objc(EXUpdatesUpdateError)
 public enum UpdateError: Int, Error {
   case invalidExpoProtocolVersion
+  case legacyManifestInstantiationInvalid
 }
 
 @objc(EXUpdatesUpdate)
@@ -138,11 +139,7 @@ public class Update: NSObject {
     let protocolVersion = responseHeaderData.protocolVersion
     switch protocolVersion {
     case nil:
-      return LegacyUpdate.update(
-        withLegacyManifest: LegacyManifest(rawManifestJSON: withManifest),
-        config: config,
-        database: database
-      )
+      throw UpdateError.legacyManifestInstantiationInvalid
     case 0, 1:
       return NewUpdate.update(
         withNewManifest: NewManifest(rawManifestJSON: withManifest),
