+++ packages/expo-updates/ios/Tests/FileDownloaderManifestParsingSpec.swift
@@ -28,11 +28,11 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
           url: URL(string: "https://exp.host/@test/test")!,
           statusCode: 200,
           httpVersion: "HTTP/1.1",
-          headerFields: ["content-type": contentType]
+          headerFields: ["content-type": contentType, "expo-protocol-version": "0"]
         )!
         
-        let bodyData = CertificateFixtures.testClassicManifestBody.data(using: .utf8)!
-        
+        let bodyData = CertificateFixtures.testNewManifestBody.data(using: .utf8)!
+
         var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
         downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
@@ -59,12 +59,12 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
           url: URL(string: "https://exp.host/@test/test")!,
           statusCode: 200,
           httpVersion: "HTTP/1.1",
-          headerFields: ["content-type": contentType]
+          headerFields: ["content-type": contentType, "expo-protocol-version": "0"]
         )!
         
         let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
           boundary: boundary,
-          manifest: CertificateFixtures.testClassicManifestBody,
+          manifest: CertificateFixtures.testNewManifestBody,
           manifestSignature: nil,
           certificateChain: nil,
           directive: CertificateFixtures.testDirectiveNoUpdateAvailable,
